fix: profile DTO converter was reading [Object object]

+++ src/main/java/fr/siamois/domain/services/ApplicationMembersServiceInterfaceImpl.java
@@ -2,6 +2,7 @@ package fr.siamois.domain.services;
 
 import fr.siamois.domain.models.auth.Person;
 import fr.siamois.domain.models.permissions.PersonProfileAssignment;
+import fr.siamois.domain.models.permissions.ProfileConstants;
 import fr.siamois.domain.services.permissions.PersonProfileAssignmentService;
 import fr.siamois.domain.services.permissions.ProfileService;
 import fr.siamois.dto.entity.ApplicationMemberDTO;
@@ -64,7 +65,12 @@ public class ApplicationMembersServiceInterfaceImpl implements ApplicationMember
     }
 
     @Override
-    public void removeProfileFromMember(ApplicationMemberDTO member, ProfileDTO profile) {
+    public boolean removeProfileFromMember(ApplicationMemberDTO member, ProfileDTO profile) {
+        if (ProfileConstants.SUPERADMIN.equals(profile.getCode())
+                && !personProfileAssignmentService.isNotLastSuperAdmin(member.getPerson())) {
+            return false;
+        }
         personProfileAssignmentService.remove(member.getPerson(), profile);
+        return true;
     }
 }
